@@ -1,27 +1,20 @@
 {
-  "proxy": "http://localhost:3000",
-  "homepage": ".",
-  "name": "udemyinstafe",
+  "name": "miniinsta",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.5.1",
-    "@mui/material": "^5.5.3",
-    "@mui/styled-engine-sc": "^5.5.2",
-    "@mui/styles": "^5.5.3",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-scripts": "5.0.0",
-    "styled-components": "^5.3.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "web-vitals": "^2.1.4"
   },
+  "devDependencies": {
+    "react-scripts": "5.0.1"
+  },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "PORT=3000 react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
